Add main screen with dynamic content area.
Login screen refactored to be a 'page' that can be displayed in the main
windows dynamic canvas.

Main screen has a permanent navigation bar at the top, and status bar
at the bottom.  The rest of the screen is made up from whatever page is
currently filling the dynamic canvas.