Add basic auth token persistence
This commit adds a first pass at saving & loading the auth token.  It is
likely to change when the AniList client is created as that is the only
consumer of the token.  But for now this demonstrates the ability to
save the token so the user does not need to login every time Hisame
starts up.